translateMemberIds function moved to the core util
