+++ src/features/HierarchyViewer/components/CustomLayout/CirclePackingPanel.tsx
@@ -2,7 +2,7 @@ import { Tooltip } from '@mui/material'
 import * as d3Hierarchy from 'd3-hierarchy'
 import * as d3Selection from 'd3-selection'
 import * as d3Zoom from 'd3-zoom'
-import { useEffect, useRef, useState } from 'react'
+import { useContext, useEffect, useRef, useState } from 'react'
 import { Network } from '../../../../models/NetworkModel'
 import {
   CirclePackingType,
@@ -32,6 +32,7 @@ import { SearchState } from '../../../../models/FilterModel/SearchState'
 import { useFilterStore } from '../../../../store/FilterStore'
 import { useRendererFunctionStore } from '../../../../store/RendererFunctionStore'
 import { useCredentialStore } from '../../../../store/CredentialStore'
+import { AppConfigContext } from '../../../../AppConfigContext'
 
 interface CirclePackingPanelProps {
   rendererId: string
@@ -57,6 +58,8 @@ export const CirclePackingPanel = ({
   initialSize,
   visible,
 }: CirclePackingPanelProps): JSX.Element => {
+  const { ndexBaseUrl } = useContext(AppConfigContext)
+
   // Use this ref to access the SVG element generated by D3
   const ref = useRef<SVGSVGElement>(null)
 
@@ -500,33 +503,37 @@ export const CirclePackingPanel = ({
       circlePackingView?.hierarchy as d3Hierarchy.HierarchyNode<D3TreeNode>
 
     if (rootNode === undefined) {
-      createTreeLayout(network, nodeTable, getToken, rootNetworkId).then(
-        (root) => {
-          rootNode = root
-          if (rootNode && Object.keys(rootNode).length === 0) return
+      createTreeLayout({
+        network,
+        nodeTable,
+        url: ndexBaseUrl,
+        getToken,
+        rootNetworkId,
+      }).then((root) => {
+        rootNode = root
+        if (rootNode && Object.keys(rootNode).length === 0) return
 
-          const updatedView = applyVisualStyle({
-            network: network,
-            visualStyle: visualStyle,
-            nodeTable: nodeTable,
-            edgeTable: edgeTable,
-            networkView: primaryView,
-          })
+        const updatedView = applyVisualStyle({
+          network: network,
+          visualStyle: visualStyle,
+          nodeTable: nodeTable,
+          edgeTable: edgeTable,
+          networkView: primaryView,
+        })
 
-          // Create a new Circle Packing view model
-          const width = initialSize?.w ?? 0
-          const height = initialSize?.h ?? 0
-          const cpViewModel: CirclePackingView = createCirclePackingView(
-            updatedView,
-            rootNode,
-            width,
-            height,
-          )
+        // Create a new Circle Packing view model
+        const width = initialSize?.w ?? 0
+        const height = initialSize?.h ?? 0
+        const cpViewModel: CirclePackingView = createCirclePackingView(
+          updatedView,
+          rootNode,
+          width,
+          height,
+        )
 
-          // Register the new view model
-          addViewModel(network.id, cpViewModel)
-        },
-      )
+        // Register the new view model
+        addViewModel(network.id, cpViewModel)
+      })
     }
   }
